continue refactoring & testing clusterOI
