+++ ui-tests/clusterOI.spec.js
@@ -81,11 +81,13 @@ const previewClusterOI = async (page) => {
  * Assumes that the clusterOI editor has nodes
  * Assumes that a current clusterOI does not have the same name
  */
-const saveClusterOI = async (page, name, expectDialog = false) => {
+const saveClusterOI = async (page, name, expectDialog = false, newCluster = true) => {
   await page.locator("#priority-panel-save").click();
-  await expect(page.locator(".has-error")).toBeVisible();
-
-  await page.locator('[data-hivtrace-ui-role="priority-panel-name"]').fill(name);
+  if (newCluster) {
+    // user gets prompted to enter cluster name
+    await expect(page.locator(".has-error")).toBeVisible();
+    await page.locator('[data-hivtrace-ui-role="priority-panel-name"]').fill(name);
+  }
 
   if (expectDialog) {
     const acceptDialog = getAcceptDialogFunction("This cluster of interest does not include all the nodes in the current");
@@ -134,8 +136,6 @@ test('clusterOI editor opens, can add nodes', async ({ page }) => {
 });
 
 test('preview cluster and then open clusterOI editor', async ({ page }) => {
-  await page.goto('http://127.0.0.1:8080/html/priority-sets-args.html?network=../ui-tests/data/network.json');
-
   await page.locator(".cluster-group").first().click();
   await page.getByText("Show this cluster in separate tab", { exact: true }).click();
   await expect(page.locator("#top_level_tab_container")
@@ -202,4 +202,103 @@ test('add nodes via graph to clusterOI editor, save, clone clusterOI, save, and 
 
   await expect(await page.locator("#priority_set_table tbody tr"))
     .toHaveCount(0);
+})
+
+test('create clusterOI and save, then view history', async ({ page }) => {
+  await openEditor(page);
+
+  await page.locator("#trace-default-tab").click();
+  await page.locator(".cluster-group").first().click();
+  await page.getByText("Add this cluster to the cluster of interest", { exact: true }).click();
+
+  // check that the nodes are added, empty clusterOI text shouldn't exist 
+  await expect(page.getByText("clusterOI editor (0 nodes)")).toHaveCount(0);
+
+  await previewClusterOI(page);
+
+  await saveClusterOI(page, "Cluster 1");
+
+  // view history
+  await page.locator(".view-edit-cluster").first().click();
+  await page.getByText("View history over time", { exact: true }).click();
+
+  await expect(page.getByText("History of Cluster 1")).toBeVisible();
+
+  await expect(page.locator("div.tab-pane.active")
+    .filter({ has: page.locator("svg") }) // history over time graph
+    .filter({ has: page.getByText("BMK383478US2007") })
+    .filter({ has: page.getByText("BMK383725US2014") }))
+    .toBeVisible();
+})
+
+test('create clusterOI, view nodes list, edit, and view nodes list again', async ({ page }) => {
+  await openEditor(page);
+
+  await page.locator("#trace-default-tab").click();
+  await page.locator(".cluster-group").first().click();
+  await page.getByText("Add this cluster to the cluster of interest", { exact: true }).click();
+
+  // check that the nodes are added, empty clusterOI text shouldn't exist 
+  await expect(page.getByText("clusterOI editor (0 nodes)")).toHaveCount(0);
+
+  await previewClusterOI(page);
+
+  await saveClusterOI(page, "Cluster 1");
+
+  // view nodes list
+  await page.locator(".view-edit-cluster").first().click();
+  await page.getByText("View nodes in this cluster of interest", { exact: true }).click();
+
+  await page.locator("[data-hivtrace-ui-role='cluster_list_view_toggle']").click();
+  await expect(page.locator("[data-hivtrace-ui-role='cluster_list_body']")
+    .filter({ has: page.getByText("BMK384750US2015") }))
+    .toBeVisible();
+
+  await expect(page.getByText("BMK384750US2015")).toHaveCount(3); // Two in the node view, one hidden in the nodes table
+
+  // edit nodes
+  await page.locator("[data-hivtrace-ui-role='cluster_list_body']").getByText("Dismiss", { exact: true }).click();
+  await page.locator(".view-edit-cluster").first().click();
+  await page.getByText("Modify this cluster of interest", { exact: true }).click();
+
+  await page.locator("#priority-panel-node-table")
+    .locator("tr")
+    .filter({ has: page.getByText("BMK384750US2015") })
+    .locator("i.fa.fa-trash")
+    .click();
+
+  // TODO: determine why the page is sometimes getting redirected (never happens during manual testing)
+  // wait two seconds because of unpredictable delete button page redirection behavior  
+  await page.waitForTimeout(2000);
+
+  await page.getByRole('button', { name: 'Delete' }).click();
+
+  await expect(page.locator(".jsPanel")
+    .first()
+    .filter({ has: page.getByText("BMK384750US2015") }))
+    .toHaveCount(0);
+
+  await expect(page.locator(".jsPanel")
+    .first()
+    .filter({ has: page.getByText("BMK384168US2005") }))
+    .toHaveCount(1);
+
+  await saveClusterOI(page, "Cluster 1", false, false);
+
+  // expect node size to be 42
+  await expect(page.locator("#priority_set_table")
+    .filter({ has: page.getByText("Cluster 1") })
+    .filter({ has: page.getByText("42") }))
+    .toBeVisible();
+
+  // view nodes list again and check that the node is not there
+  await page.locator(".view-edit-cluster").first().click();
+  await page.getByText("View nodes in this cluster of interest", { exact: true }).click();
+
+  await expect(page.locator("[data-hivtrace-ui-role='cluster_list_body']")
+    .filter({ has: page.getByText("BMK384750US2015") }))
+    .toHaveCount(0);
+
+  await expect(page.getByText("BMK384750US2015")).toHaveCount(1); // Only the one hidden in the nodes table
+  await expect(page.getByText("BMK384168US2005")).toHaveCount(3); // Two in the node view, one hidden in the nodes table
 })
